Reformat

+++ SIR.java
@@ -314,7 +314,7 @@ public class SIR {
         numero *= 10;
         numero= Math.floor(numero);
         numero /= 10;
-        return numero % 1 == 0;
+        return (float) numero % 1 == 0;
     }
 
     public static void escreverPontosGnu(double[] parametro, int numeroDeDias, String ficheiroGNU, double h) throws FileNotFoundException {
